+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to the Create Flutter Widgets and Classes extension will be 
 
 ## [1.2.0]
 
+- Added the code action `Implements interface`.
 - Added the code action `Wrap with Hero`.
 - Added the code action `Wrap with Obx`.
 - Added the code action `Wrap with SizedBox`.
+- Removed the setting `fileNameCaseFormat`.
 
 ## [1.1.4]